moved a struct
